control-plane/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexusml"
version = "0.2.0"
description = "End-to-end ML versioning and serving platform - Git-integrated model storage with high-performance inference"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Samir Varma"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "gitpython>=3.1.40",
    "boto3>=1.28.0",
    "google-cloud-storage>=2.10.0",
    "pyyaml>=6.0",
]

[project.scripts]
nexus = "nexus.cli:app"

[tool.setuptools]
packages = ["nexus"]
